break out NewThresholdField
