+++ api/serializers/__init__.py
@@ -1,3 +1,4 @@
 from .get_context_user import get_context_user
 from .get_projects_for_obj import get_projects_for_obj
 from .projects_field import ProjectsField
+from .new_threshold_field import NewThresholdField
